+++ features/stripe-connect/components/status-views/UnverifiedView.tsx
@@ -7,9 +7,11 @@
 
 import Link from "next/link";
 
-import { ArrowRight, CircleAlert } from "lucide-react";
+import { ArrowRight, CircleAlert, Lock } from "lucide-react";
 
 import { Button } from "@/components/ui/button";
+
+import { OnboardingIntro } from "../OnboardingIntro";
 
 interface UnverifiedViewProps {
   refreshUrl: string;
@@ -17,14 +19,16 @@ interface UnverifiedViewProps {
 
 export function UnverifiedView({ refreshUrl }: UnverifiedViewProps) {
   return (
-    <div className="space-y-4">
+    <div>
+      <OnboardingIntro hasExistingAccount />
+
       <div className="rounded-xl border border-amber-500/25 bg-amber-500/5 p-4">
         <div className="flex gap-3 items-start">
           <div className="shrink-0 rounded-lg bg-amber-500/15 p-2 flex items-center justify-center">
             <CircleAlert className="h-4 w-4 text-amber-600" />
           </div>
           <div className="min-w-0">
-            <p className="text-sm font-semibold">設定を開始してください</p>
+            <p className="text-sm font-semibold">設定を再開しましょう</p>
             <p className="mt-1 text-xs leading-relaxed text-muted-foreground">
               オンライン集金設定が完了していません。
             </p>
@@ -34,14 +38,19 @@ export function UnverifiedView({ refreshUrl }: UnverifiedViewProps) {
 
       <Button
         asChild
-        className="group relative h-11 w-full rounded-xl border border-amber-500/25 bg-gradient-to-r from-amber-500/10 via-amber-500/5 to-transparent text-sm font-semibold text-amber-800 transition-all duration-300 hover:border-amber-500/40 hover:from-amber-500/20 hover:via-amber-500/10 hover:to-amber-500/5 hover:text-amber-900 shadow-[inset_0_1px_0_hsl(var(--background)/0.4),0_8px_16px_-12px_rgba(245,158,11,0.2)] hover:shadow-[inset_0_1px_0_hsl(var(--background)/0.5),0_12px_24px_-12px_rgba(245,158,11,0.4)]"
-        variant="outline"
+        className="w-full mt-6 h-12 text-base font-semibold shadow-md hover:shadow-lg transition-shadow"
+        size="lg"
       >
         <Link href={refreshUrl} prefetch={false}>
-          Stripeで設定を始める
-          <ArrowRight className="ml-2 h-4 w-4 transition-transform group-hover:translate-x-1" />
+          設定を再開する
+          <ArrowRight className="ml-2 h-5 w-5" />
         </Link>
       </Button>
+
+      <p className="flex items-center justify-center gap-1.5 text-xs text-muted-foreground mt-3">
+        <Lock className="h-3 w-3" />
+        Stripeの安全な画面で設定します・約3分で完了
+      </p>
     </div>
   );
 }
